It's fun.

+++ inc/RandomEngine.h
@@ -3,5 +3,5 @@
 // 初始化随机数生成器。
 void InitializeRandomEngine();
 
-// 生成下一个随机数。
-double NextRand();
+// 生成下一个随机数。(在0~1023范围内。)
+int NextRand();
